Extract trailing macro semicolon handling to semicolon-fixes.ts
Moves removeTrailingMacroSemicolons and related helpers from the monolithic
apply-feather-fixes.ts (15,676 lines) into the existing semicolon-fixes.ts
module where they belong architecturally. This reduces the main file by 115
lines and consolidates all semicolon-related Feather fixes into a single,
focused module.

Changes:
- Move TRAILING_MACRO_SEMICOLON_PATTERN constant to semicolon-fixes.ts
- Move removeTrailingMacroSemicolons() function
- Move sanitizeMacroDeclaration() helper
- Move registerSanitizedMacroName() helper
- Update import in apply-feather-fixes.ts
- Update test import to use new location
- Add comprehensive TSDoc comments to extracted functions

No behavioral changes; all tests pass.
