+++ src/plugin/test/transforms/trailing-macro-semicolon.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import { performance } from "node:perf_hooks";
 import { describe, it } from "node:test";
 
-import { TRAILING_MACRO_SEMICOLON_PATTERN } from "../../src/transforms/feather/apply-feather-fixes.js";
+import { TRAILING_MACRO_SEMICOLON_PATTERN } from "../../src/transforms/feather/semicolon-fixes.js";
 
 void describe("TRAILING_MACRO_SEMICOLON_PATTERN", () => {
     void it("removes trailing semicolons correctly", () => {
